requirements.txt
dotenv==0.9.9
python-dotenv==1.0.1
gpiozero==2.0
pyserial==3.5
pyaudio==0.2.14
aubio==0.4.9
numpy==1.26.4
adafruit-circuitpython-neopixel==6.3.11
board==1.0
